Fix #50, implement dispatcher pattern for BP
Split command handler implementations and dispatching operation to
separate source files.
