+++ fsw/src/bp_app.h
@@ -24,30 +24,6 @@
 #include "cfe_error.h"
 #include "bp_msg.h"
 
-/*-----------------------------------------------
- * Command Handlers
- *-----------------------------------------------*/
-CFE_Status_t BP_NoopCmd(const BP_NoopCmd_t *cmd);
-CFE_Status_t BP_ResetAppCmd(const BP_ResetAppCmd_t *cmd);
-CFE_Status_t BP_ReloadFlowTableCmd(const BP_ReloadFlowTableCmd_t *cmd);
-CFE_Status_t BP_EnableFlowCmd(const BP_EnableFlowCmd_t *cmd);
-CFE_Status_t BP_DisableFlowCmd(const BP_DisableFlowCmd_t *cmd);
-CFE_Status_t BP_FlushFlowCmd(const BP_FlushFlowCmd_t *cmd);
-CFE_Status_t BP_PauseFlowCmd(const BP_PauseFlowCmd_t *cmd);
-CFE_Status_t BP_ResumeFlowCmd(const BP_ResumeFlowCmd_t *cmd);
-CFE_Status_t BP_OverrideTimeoutCmd(const BP_OverrideTimeoutCmd_t *cmd);
-CFE_Status_t BP_DisableOverrideTimeoutCmd(const BP_DisableOverrideTimeoutCmd_t *cmd);
-CFE_Status_t BP_OverridePriorityCmd(const BP_OverridePriorityCmd_t *cmd);
-CFE_Status_t BP_DisableOverridePriorityCmd(const BP_DisableOverridePriorityCmd_t *cmd);
-
-CFE_Status_t BP_SendAppTlmCmd(const BP_SendAppTlmCmd_t *cmd);
-CFE_Status_t BP_SendEnabledFlowTlmCmd(const BP_SendEnabledFlowTlmCmd_t *cmd);
-CFE_Status_t BP_SendDisabledFlowTlmCmd(const BP_SendDisabledFlowTlmCmd_t *cmd);
-
-CFE_Status_t BP_ProcessWakeupCmd(const CFE_MSG_CommandHeader_t *cmd);
-
-CFE_Status_t BP_ProcessPkt(const CFE_SB_Buffer_t *MessagePtr);
-
 /************************************************
  * Exported Functions
  ************************************************/
